backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0
PyPDF2==3.0.1
Pillow==10.2.0
pytesseract==0.3.10
websockets==12.0
aiosqlite==0.19.0

# Speech-to-text (local Whisper)
faster-whisper==1.0.1

